refactor: move components into folders
