+++ src/components/ListOfCards/ListOfCards.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from './Card';
+import Card from '../Card/Card';
 import { IProduct } from 'interfaces';
 
 interface IListOfCardsProps {
